revised shared qsqlpsql plugin dependency handling for package qt
Previously we added support for postgresql+openssl in shared Qt by explicitly
linking to required libraries in psql.pro. However, we neglected to handle the
case of shared build of Qt with system-zlib, which requires -lz. This patch
takes care of -lz but also handles the situation more elegantly. Now we use
QT_LFLAGS_PSQL as intended. We set it in the configure script with help from
pkg-config for openssl dependencies. win32-g++ is now handled the unix way in
psql.pro. This is consistent with other zlib-related work. For non system-zlib
we remove the -lz.

Also, as of now qt-win32.patch is generated by git diff.