Add package opusfile
Opusfile is a decoder for Opus files. It provides a higher-level API
than the opus library, of which it makes use.

Update function taken directly from opus.mk. Two patches are needed to
make it work, one is taken from the upstream repository (change
introduced after the latest release), the other is by myself.