Added the Alure package.
The alure package is an conveniance library for openal that packs certain OpenAL functionality
in a os independent wrapper. It also can use certain audio libraries to extend the type of loadable
audio files. The current package setup depends on all the audio libraries that mxe currently has
and that Alure can handle.

Moved VERBOSE=1 to the end of the make line as per request.

Also moved the cmake_toolchain_file line to the end of the cmake command block.

Moved from a insource to an outsource build of alure with cmake