new package libusb1
This is a package for the libusb-1.0 shared library (MXE package name
"libusb1") used by various projects. It is not to be confused with the
older libusb-0.1 library (MXE package name "libusb") which is a
completely different library with different API. Both libs are in
wide-spread use and can be installed in parallel.