pkgconf: work around for buggy FindOpenSSL.cmake
FindOpenSSL.cmake bundled with cmake older than 2.8.12-rc1 uses
CMAKE_CROSS_COMPILING, which is incorrect because the real variable name
is CMAKE_CROSSCOMPILING. Because it is not possible to modify users'
CMake installation, work around it here.

This fixes building libwebsockets on Ubuntu 13.10. (I only tested 14.04
when merging the pull request.)

Signed-off-by: Timothy Gu <timothygu99@gmail.com>