PKG_UPDATE for jansson.mk
I noticed how the author for jansson.mk in #646 mentioned the lack for resolving the SHA for the PKG_UPDATE. This patch along with #649 fixes this problem.

A side effect from this is that PKG_VERSION will have to be replaced by a checksum so that it resolves the correct github url. I'll leave it up to you how to deal with this.