postgresql: require/hint minimum autoconf version 2.63
the original configure.in warns if autoconf version is less than 2.63, but
patch 1 removes this warning. this patch adds that requirement using AC_PREREQ

on debian systems where both old/new versions of autoconf are installed,
a script is used to determine which version to use. adding AC_PREREQ to
configure.in helps the script choose the correct version.

the system should already have such version since MXE itself requires 2.67