Shared build of libgpg_error failing
When building libgpg_error with a target of i686-w64-mingw32.shared
the build fails with undefined references. In the make process for
libgpg_error an application is built and then run called mkheader.
This application is used to build the gpg-error.h file. In this
application there is a strcmp for 'mingw32' on host_os variable
which failes to be true when host_os contains mingw32.shared or
mingw32.static. The failure of this test keep w32-add.h from being
added to gpg-error.h, hence causing the undefined references errors.
This is an attempt to fix this issue, officially issue #677.