fix libgcrypt on x86_64 (automake < 1.12)
automake < 1.12 didn't have serial-tests and gives an error if
it sees this, but for automake >= 1.13 serial-tests is
required so we have to include it.

Solution is to test for the version of automake (by running an
external command) and provide serial-tests if necessary.

See #681
See https://www.redhat.com/archives/libguestfs/2013-February/msg00102.html