New package: enet library
ENet's purpose is to provide a relatively thin, simple and robust network communication layer on top of UDP (User Datagram Protocol).The primary feature it provides is optional reliable, in-order delivery of packets. ENet omits certain higher level networking features such as authentication, lobbying, server discovery, encryption, or other similar tasks that are particularly application specific so that the library remains flexible, portable, and easily embeddable.

Patch contains an `enet-test.c` program and was compiled and tested on Debian Jessie with the following targets

 * x86_64-w64-mingw32.static
 * i686-w64-mingw32.static
 * x86_64-w64-mingw32.shared
 * i686-w64-mingw32.shared