Add package 'HDF-EOS2'.
http://hdfeos.org/software/library.php

"The HDF-EOS2 is a software library designed built on HDF4 to support
EOS-specific data structures, namely Grid, Point, and Swath. The new
data structures are constructed from standard HDF data objects, using
EOS conventions, through the use of a software library. A key feature of
HDF-EOS files is that instrument-independent services, such as
subsetting by geolocation, can be applied to the files across a wide
variety of data products."