fix races of downloading and verifying
Sometimes `make download -j 6` fails with errors in downloading
package openthreads (and some other). How to reproduce:

    $ sed 's/wget/wget --limit-rate=1M/' -i Makefile
    $ make download-only-plib download-only-openscenegraph \
        download-only-openthreads -j 2
    ...
    Download failed or wrong checksum of package openscenegraph!

In previous implementation download pipe (wget, gzip) used to write
its output directly to a target file in pkg/ directory.
If multiple packages (e.g., openscenegraph and openthreads) share
the same file in pkg/ directory, the download can result in failure:
while first package is verifying the checksum, another one overwrites
the file with partially downloaded content.

(Package plib in the example is needed to shift start of download of
openthreads. `--limit-rate=1M` is needed to increase duration of download.)

To fix this problem, wget command was changed to write to a unique
file, which is renamed to a target file afterwards. Renaming is
an atomic action.