Add meson plugin
This plugin adds the Meson build system. Cross files and wrappers are
installed for each target, similar to "cmake-conf". This allows Meson
projects to seamlessly integrate with MXE.

Example usage:

    make MXE_TARGETS=x86_64-w64-mingw32.static meson
    cd destdir
    x86_64-w64-mingw32.static-meson sourcedir
    x86_64-w64-mingw32.static-meson configure -Dfoo=bar
    x86_64-w64-mingw32.static-ninja

The host's meson and ninja executables can also be used after the
initial configuration step, given their versions are compatible.

Python 3.4 or later is required to be installed on the host system.