ghostscript: improving cross-building
First, run configure with the option `--with-drivers=ALL,display`. Upstream confirmed an error in the build process that keeps the display driver disabled when building a static library, but that driver is needed if you want to use libgs through the wrapper library libspectre. So creating a static library manually from the DLL object files isn't required anymore.

Also a massive update on the patch, mostly to enable real support for `--host` on the configure script (instead of relying on `uname`). Full support for `--build` wasn't added to keep the patch smaller and to not make things more complicated as they already are (for example configure runs itself recursively when CCAUX environment variable was set). ghostscript.mk now builds the needed auxiliary tools in a separate stage (as native binaries) and deletes the unwanted third-party library sources afterwards. This ensures those binaries will run on the build system and keeping the third-party sources on that step will ensure that configuring doesn't fail if i.e. native libjpeg development files aren't found.

Builds on all four MXE targets and the libraries are confirmed to work.