XMLPUBVAR: patch the mingw block, not the cygwin
In the patch that is now in MXE, somehow the line numbers got
messed up. The result is that the XMLPUBVAR is now fixed for
the cygwin case, not for mingw. This patch fixes that. It adds
'extern' to the __declspec(dllexport) definition of XMLPUBVAR.

This is also what is already in upstream libxml2, including in the
version that is shipped with mxe. See:
https://github.com/GNOME/libxml2/commit/1eabc31401b7b8c3b5273993778f37eeef37a055