Fix Poco only building "Basic Edition"
The build script only dowloaded the sources of Poco's "Basic Edition"
which only contains a few of its components.
This commit makes it pull the complete Poco library in order to make all
of it available under MXE.

The PageCompiler and Data/ODBC module had to be disabled though as they
don't build successfully.

Fixes #2552