Added a wxWidgets-3.1.4 plugin
This makes a recent wxWidgets version available as a plugin.

This is related to pull request 53, that proposes to update the default
wxWidgets version in mxe. Note the following differences wrt the code in
that pull request:

 * this plugin uses the usual MXE_CONFIGURE_OPTS, addressing the
   comments of @tonytheodore
 * configure complaints about unused options that result from this choice
   produce warnings, rather than errors, by passing --enable-option-checking
   to configure (without argument). This was suggested by @janvdijk (me)
   in the audit trail of pull request 53
 * it makes version 3.1.4 available, rather than 3.1.3
 * no patches are needed to build this version of wxWidgets with the
   default gcc or with gcc9 or gcc10 (available as plugins in mxe).

This patch allows us to do in-the-wild testing of wxWidgets, before
making 3.1.4 the defult. (I think that would be a good idea, even soon,
given the facts that 3.0.x does not see much maintenance (if at all),
and that 3.1.x releases have proven to be VERY stable and reliable.