VTK: Fix dll import/export of vtkInformation.h
Avoid error: function definition XXX is marked dllimport errors with gcc-11,
because of the usage of VTK_INFORMATION_EXPORT, rather than VTKCOMMONCORE_EXPORT.
Somehow these problems do not manifest themselves with the stock gcc or with
the gcc10 plugin.
This patch makes the code look like the git/master version, where the
export is handled on the class level. The upstream patch is
https://github.com/Kitware/VTK/commit/8e7d271ffc6fc4e26dc22005146951e726547cc7