hidapi: Switch source to new GitHub repo and update
The README at https://github.com/libusb/hidapi says "HIDAPI library was
originally developed by Alan Ott (signal11). It was moved to
libusb/hidapi on June 4th, 2019, in order to merge important bugfixes
and continue development of the library." As such, the libusb version is
now the official repository and we should use it.

Also update to the latest master. Among other things, this fixes a
[bug][autoconf-fix] that broke the build on new versions of autoconf
(2.70 and later).

[autoconf-fix]: https://github.com/libusb/hidapi/pull/226