patch-tool-mxe: provide message when calling "git tag"
`git tag something` will invoke the editor. As the editor is invoked
via `make`, it doesn't have a TTY to display stuff on. In the case of
`vim`, this means `vim` hangs in the background indefinitely, as does
`make` waiting for `vim` to finish. The user has no way to interact
with that editor instance, having to kill it, which in turn causes
`git tag` and ultimately `make` to fail.

Fixes #2739.