qt6-conf: work around problem with cmake argument order
Qt's own private cmake is able to configure the project, but fails
with --build due to presence of other arguments. The standard cmake
for the target has a wrapper to avoid inserting unneeded
arguments with --build. Also, --builld must be the first argument.
The wrapper uses the logic in src/cmake/conf/target-cmake.in

Earlier cmake versions might have been more tolerant.