Allow separating ccache's cache dir to a different
In light of what is discussed in #2462, it seems reasonable to allow
having `ccache` put its cache to a persistent storage that is shared
through pipeline runs.  The current setup doesn't do that since it puts
the cache in `MXE_CCACHE_DIR`, a directory in which also binaries and
other files reside that should not be shared as such.