Fix wine-fail workaround
The workaround to make sure `wine` fails in packages' build processes
was to set `WINEPREFIX` to a read-only directory.  However, this fails
if this read-only directory is in fact not read-only, as can happe when
compiling MXE as `root` in a docker environment.

Setting `WINEPREFIX` to `/dev/null` fixes this without the workaround.

This closes #2819.