Avoid python version detection when not installed.
When "python" is not installed, PYTHON is empty and then the (eager) evaluation of PY_XY_VER causes an error from the shell. This silences such error by not attempting to detect the version of "python" when it is not installed.

With "bash", the error is "bash: - : invalid option", but it is not immediately clear that in fact it is benign (when python is not needed).

On some systems, "python" is not installed (but python3 / PYTHON3 is).