Fix compile error when using clang
CLang issues an error when you attempt to pass a _Noreturn function to a
function pointer marked with __attribute_noreturn__, whereas GCC only
issues a warning. As far as I can tell this doesn't actually affect
anything under the hood, and there's multiple other projects that already
apply this patch when using clang (like Gentoo). For more information
here's a thread on the GNU mailing list about the issue:

https://lists.gnu.org/archive/html/bug-gnulib/2017-04/msg00104.html

Note that the affected obstack.c file seems to be copied into several
different projects, so while I happened to encounter it in m4 and
gettext, it's probably in other packages too.