ocaml-native: add patches to fix build failure
There were two separate things causing the build failure here.

One was a file-scope global variable not being declared as extern in a header,
causing it to be defined multiple times, once in each file that included the header.
The linker wasn't happy about this, so the solution was to move the non-extern
definition to a source code file and make the header declaration extern.

The other issue was that SIGSTKSZ is no longer considered constant for the purposes
of dimensioning a file-scope array, thus necesitating its replacement with something
that is actually constant.  32k was an arbitrary choice, but the default on my current
linux machine was 8k, so this should be more than enough.