mariadb-connector-c: workaround for linking to 32-bit shared libmaria.a
For 32-bit shared mingw builds, linking to libmariadb.a seems to require
omitting __stdcall in function declarations. Otherwise, the decorated
function names do not match the names in libmariadb.a.

For example, linking to libmariadb.a fails with:
    undefined reference to `mysql_server_init@12'

while libmariadb.a provides:
    00000000 I __imp__mysql_server_init
    00000000 T _mysql_server_init

For 64-bit shared mingw builds, this workaround is not required but
harmless.

https://github.com/mxe/mxe/issues/3219