do not force to Release type successive configs
When configuration is done in multiple runs, such as
when an initial run is done to populate the options
and then succesive runs to apply the options, do not
force Release type on second run.

qtcreator does exactly that: specify build type only
on first run and then relies on cached variable.